requirements.txt
streamlit==1.18.1
numpy==1.23.5
tensorflow==2.10.1
Pillow==9.3.0

